backend/requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-cors
